Split accordion styles into separate files.
