+++ components/accordion/accordion-header.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import * as Styled from './styled';
+import * as Styled from './styled-header';
 
 export function AccordionHeader({ children }) {
 	return <Styled.AccordionHeader>{children}</Styled.AccordionHeader>;
 }
 
 AccordionHeader.propTypes = {
-	/** Children */
 	children: PropTypes.node,
 };
