+++ components/accordion/accordion-item.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import * as Styled from './styled';
+import * as Styled from './styled-item';
 
 export function AccordionItem({ children }) {
 	return <Styled.AccordionItem>{children}</Styled.AccordionItem>;
 }
 
 AccordionItem.propTypes = {
-	/** Children */
 	children: PropTypes.node,
 };
