+++ components/accordion/accordion-panel.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import * as Styled from './styled';
+import * as Styled from './styled-panel';
 
 export function AccordionPanel({ children }) {
 	return <Styled.AccordionPanel>{children}</Styled.AccordionPanel>;
 }
 
 AccordionPanel.propTypes = {
-	/** Children */
 	children: PropTypes.node,
 };
