+++ components/accordion/styled-accordion.jsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+import { resetStyles } from '../utils';
+
+export const Accordion = styled.div`
+	${resetStyles};
+`;
