+++ components/accordion/styled-panel.jsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import { resetStyles } from '../utils';
+
+export const AccordionPanel = styled.div`
+	${resetStyles};
+
+	grid-area: panel;
+`;
